requirements.txt
asttokens==2.0.5
attrs==22.1.0
backcall==0.2.0
bleach==5.0.1
click==8.1.3
colorama==0.4.5
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.2.1
entrypoints==0.4
executing==0.9.1
Faker==0.8.13
fastjsonschema==2.16.1
Flask==2.2.0
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
ipykernel==6.10.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.9.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab-widgets==1.1.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==5.7.13
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
prometheus-client==0.14.1
prompt-toolkit==2.0.10
psutil==5.9.1
pure-eval==0.2.2
Pygments==2.12.0
pymongo==4.2.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.6
pywinpty==2.0.6
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
Send2Trash==1.8.0
six==1.16.0
stack-data==0.3.0
terminado==0.15.0
testpath==0.6.0
text-unidecode==1.2
tornado==6.2
traitlets==5.3.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.1
widgetsnbextension==3.6.1
